@@ -1,14 +1,14 @@
 module github.com/cmmoran/enumer
 
+go 1.24.7
+
 require (
 	github.com/pascaldekloe/name v1.0.1
-	golang.org/x/text v0.16.0
-	golang.org/x/tools v0.21.1-0.20240508182429-e35e4ccd0d2d
+	golang.org/x/text v0.29.0
+	golang.org/x/tools v0.37.0
 )
 
 require (
-	golang.org/x/mod v0.17.0 // indirect
-	golang.org/x/sync v0.7.0 // indirect
+	golang.org/x/mod v0.28.0 // indirect
+	golang.org/x/sync v0.17.0 // indirect
 )
-
-go 1.21
