@@ -10,7 +10,7 @@
 package main
 
 import (
-	"io/ioutil"
+	"io"
 	"os"
 	"path/filepath"
 	"strings"
@@ -58,6 +58,10 @@ var goldenJSONAndSQL = []Golden{
 
 var goldenTrimPrefix = []Golden{
 	{"trimPrefix", trimPrefixIn},
+}
+
+var goldenTrimSuffix = []Golden{
+	{"trimSuffix", trimSuffixIn},
 }
 
 var goldenTrimPrefixMultiple = []Golden{
@@ -289,6 +293,18 @@ const (
 )
 `
 
+const trimSuffixIn = `type Day int
+const (
+	MondayDay Day = iota
+	TuesdayDay
+	WednesdayDay
+	ThursdayDay
+	FridayDay
+	SaturdayDay
+	SundayDay
+)
+`
+
 const trimPrefixMultipleIn = `type Day int
 const (
 	DayMonday Day = iota
@@ -315,52 +331,54 @@ const (
 
 func TestGolden(t *testing.T) {
 	for _, test := range golden {
-		runGoldenTest(t, test, false, false, false, false, false, false, true, "", "")
+		runGoldenTest(t, test, false, false, false, false, false, false, true, "", "", "", "")
 	}
 	for _, test := range goldenJSON {
-		runGoldenTest(t, test, true, false, false, false, false, false, false, "", "")
+		runGoldenTest(t, test, true, false, false, false, false, false, false, "", "", "", "")
 	}
 	for _, test := range goldenText {
-		runGoldenTest(t, test, false, false, false, true, false, false, false, "", "")
+		runGoldenTest(t, test, false, false, false, true, false, false, false, "", "", "", "")
 	}
 	for _, test := range goldenYAML {
-		runGoldenTest(t, test, false, true, false, false, false, false, false, "", "")
+		runGoldenTest(t, test, false, true, false, false, false, false, false, "", "", "", "")
 	}
 	for _, test := range goldenSQL {
-		runGoldenTest(t, test, false, false, true, false, false, false, false, "", "")
+		runGoldenTest(t, test, false, false, true, false, false, false, false, "", "", "", "")
 	}
 	for _, test := range goldenJSONAndSQL {
-		runGoldenTest(t, test, true, false, true, false, false, false, false, "", "")
+		runGoldenTest(t, test, true, false, true, false, false, false, false, "", "", "", "")
 	}
 	for _, test := range goldenGQLGen {
-		runGoldenTest(t, test, false, false, false, false, false, true, false, "", "")
+		runGoldenTest(t, test, false, false, false, false, false, true, false, "", "", "", "")
 	}
 	for _, test := range goldenTrimPrefix {
-		runGoldenTest(t, test, false, false, false, false, false, false, false, "Day", "")
+		runGoldenTest(t, test, false, false, false, false, false, false, false, "Day", "", "", "")
+	}
+	t.Logf("trimSuffix")
+	for _, test := range goldenTrimSuffix {
+		runGoldenTest(t, test, false, false, false, false, false, false, false, "", "Day", "", "")
 	}
 	for _, test := range goldenTrimPrefixMultiple {
-		runGoldenTest(t, test, false, false, false, false, false, false, false, "Day,Night", "")
+		runGoldenTest(t, test, false, false, false, false, false, false, false, "Day,Night", "", "", "")
 	}
 	for _, test := range goldenWithPrefix {
-		runGoldenTest(t, test, false, false, false, false, false, false, false, "", "Day")
+		runGoldenTest(t, test, false, false, false, false, false, false, false, "", "", "Day", "")
 	}
 	for _, test := range goldenTrimAndAddPrefix {
-		runGoldenTest(t, test, false, false, false, false, false, false, false, "Day", "Night")
+		runGoldenTest(t, test, false, false, false, false, false, false, false, "Day", "", "Night", "")
 	}
 	for _, test := range goldenLinecomment {
-		runGoldenTest(t, test, false, false, false, false, true, false, false, "", "")
+		runGoldenTest(t, test, false, false, false, false, true, false, false, "", "", "", "")
 	}
 }
 
-func runGoldenTest(t *testing.T, test Golden,
-	generateJSON, generateYAML, generateSQL, generateText, linecomment, generateGQLGen, generateValuesMethod bool,
-	trimPrefix string, prefix string) {
+func runGoldenTest(t *testing.T, test Golden, generateJSON, generateYAML, generateSQL, generateText, linecomment, generateGQLGen, generateValuesMethod bool, trimPrefix, trimSuffix, prefix, suffix string) {
 
 	var g Generator
 	file := test.name + ".go"
 	input := "package test\n" + test.input
 
-	dir, err := ioutil.TempDir("", "stringer")
+	dir, err := os.MkdirTemp("", "stringer")
 	if err != nil {
 		t.Error(err)
 	}
@@ -372,7 +390,7 @@ func runGoldenTest(t *testing.T, test Golden,
 	}()
 
 	absFile := filepath.Join(dir, file)
-	err = ioutil.WriteFile(absFile, []byte(input), 0644)
+	err = os.WriteFile(absFile, []byte(input), 0644)
 	if err != nil {
 		t.Error(err)
 	}
@@ -382,7 +400,7 @@ func runGoldenTest(t *testing.T, test Golden,
 	if len(tokens) != 3 {
 		t.Fatalf("%s: need type declaration on first line", test.name)
 	}
-	g.generate(tokens[1], generateJSON, generateYAML, generateSQL, generateText, generateGQLGen, "noop", trimPrefix, prefix, linecomment, generateValuesMethod)
+	g.generate(tokens[1], generateJSON, generateYAML, generateSQL, generateText, generateGQLGen, "noop", trimPrefix, trimSuffix, prefix, suffix, linecomment, generateValuesMethod)
 	got := string(g.format())
 	if got != loadGolden(test.name) {
 		// Use this to help build a golden text when changes are needed
@@ -401,7 +419,7 @@ func loadGolden(name string) string {
 		return ""
 	}
 	defer fh.Close()
-	b, err := ioutil.ReadAll(fh)
+	b, err := io.ReadAll(fh)
 	if err != nil {
 		return ""
 	}
